src/test/java/com/leetcode/backtracking/hard/NQueensII_52Test.java
package com.leetcode.backtracking.hard;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class NQueensII_52Test {
  private final NQueensII_52 solution = new NQueensII_52();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(4, 2), arguments(1, 1));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void totalNQueens(int n, int expectedResult) {
    int actualResult = solution.totalNQueens(n);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
